requirements.txt
discord.py
discord.py[voice]
aiohttp
openai
pyyaml
google-genai
PyNaCl
yt-dlp
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
aiofiles
matplotlib
japanize-matplotlib
cartopy
langdetect
pillow
#セルフホストがめんどくさい人向けのリリース用
nuitka